gateway/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "gateway"
version = "0.1.0"
description = "Gateway service for the Chicago Vibe Coding Workshop"
authors = [{name = "Developer", email = "dev@example.com"}]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "databricks-zerobus",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.1.1",
    "asyncpg>=0.29.0",
    "uuid>=1.30",
]

[project.scripts]
gateway = "gateway.main:main"
gateway-ingest-example = "gateway.ingest.example:main"

[tool.hatch.build.targets.wheel]
packages = ["src/gateway"]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.25.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.11"
strict = true

[tool.uv.sources]
databricks-zerobus = { path = "vendor/databricks_zerobus-0.0.15-py3-none-any.whl" }
